fix(wiki): reserve the chat side panel's width so doc text stops at its edge
The chat side panel is a viewport-edge overlay and nothing shrank the area it
covers, so every line of a document ran underneath it and was clipped mid-word.
FileView cannot gate on this itself: the panel's open state lives in
WikiToolbar, a descendant, which is why the existing rail-reserved and
panel-anchored reservations never applied to it.

The panel now publishes its measured width and the doc area reserves it, so
every centred block re-centres inside what stays visible and the container
queries resolve against the reduced box. Closing the panel releases it.

This generalises the height publisher added for the bottom dock into one
usePublishedSize hook over an axis, rather than growing a second near-identical
copy of measure, observe and clean up.

Only the doc surface reserves. The folder listing hosts the same panel behind a
different scrim layering and is left alone deliberately.
