+++ frontend/src/app/globals.css
@@ -493,6 +493,14 @@ body {
   padding-bottom: calc(var(--wiki-dock-height, 0px) + 1.5rem);
 }
 
+/* The chat side panel is a viewport-edge overlay, so the area it covers has to
+   shrink or the doc's lines run underneath it. Reserving the width the panel
+   publishes re-centres every block inside what stays visible, and the container
+   queries below resolve against the reduced box. */
+.chat-panel-reserved {
+  padding-right: var(--wiki-chat-panel-width, 0px);
+}
+
 /* Anchored comments panel: the editor's native scrollbar would sit at the
    doc/panel boundary, so it hides in favor of the panel's viewport-edge
    scrollbar (EditorEdgeScrollbar). */
