+++ frontend/src/components/wiki/toolbar/chatParts.tsx
@@ -2,7 +2,7 @@
 
 // Shared visual pieces for wiki toolbar chrome, modes, responses, and input.
 // Layout and styling follow the toolbar mock tokens.
-import { useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 
 import {
   Button,
@@ -47,6 +47,37 @@ import type { ChatFeedback } from "@/lib/chat";
 import { toast } from "@/hooks/useToast";
 import { updateUserSettings, useUserSettings } from "@/lib/userSettings";
 import { useWikiTree } from "@/lib/wiki/hooks";
+
+/** Publishes an element's measured size as a root CSS property, so surfaces
+ *  that cannot see this component's state still reserve room for it. Both chat
+ *  overlays open from state owned below the layout they cover. */
+export function usePublishedSize(
+  property: string,
+  axis: "height" | "width",
+  active = true,
+) {
+  const ref = useRef<HTMLDivElement>(null);
+  // Callers republish on every render, since these overlays resize for reasons
+  // no dependency list spans. The observer catches what no render reports.
+  const publish = useCallback(() => {
+    const el = ref.current;
+    if (!active || !el) return;
+    const px = axis === "height" ? el.offsetHeight : el.offsetWidth;
+    document.documentElement.style.setProperty(property, `${px}px`);
+  }, [property, axis, active]);
+  useEffect(() => {
+    const el = ref.current;
+    if (!active || !el) return;
+    publish();
+    const ro = new ResizeObserver(publish);
+    ro.observe(el);
+    return () => {
+      ro.disconnect();
+      document.documentElement.style.removeProperty(property);
+    };
+  }, [property, active, publish]);
+  return { ref, publish };
+}
 
 export type ToolbarMode = "chat" | "watch" | "launch";
